+++ src/readme.spec.ts
@@ -380,16 +380,16 @@ describe('README examples', () => {
           }
 
           @Validate({ type: 'boolean' })
-          aBoolean = false as boolean;
+          public aBoolean = false as boolean;
 
           @Validate({ type: 'number' })
-          aNumber = 123 as number;
+          public aNumber = 123 as number;
 
           @Validate({ type: 'string' })
-          aString = 'abc' as string;
+          public aString = 'abc' as string;
 
           @Validate({ type: 'object' })
-          anObject = {} as object;
+          public anObject = {} as object;
         }
 
         const instance = new MyValidClass();
@@ -403,10 +403,10 @@ describe('README examples', () => {
 
         assert(instance.valid === false);
         assert(Object.keys(instance.errors).length === 4);
-        assert(instance.errors['aBoolean'].length === 1);
-        assert(instance.errors['aNumber'].length === 1);
-        assert(instance.errors['aString'].length === 1);
-        assert(instance.errors['anObject'].length === 1);
+        assert(instance.errors.aBoolean.length === 1);
+        assert(instance.errors.aNumber.length === 1);
+        assert(instance.errors.aString.length === 1);
+        assert(instance.errors.anObject.length === 1);
       });
     });
 
@@ -421,7 +421,7 @@ describe('README examples', () => {
 
             // Validate year is within a predefined scope
             @Validate({ provider: (value, obj) => value > 1900 && value < 2100 })
-            born = undefined as number;
+            public born = undefined as number;
 
             // Validate year is within dynamic scope and throw custom validation errors
             @Validate({
@@ -448,23 +448,23 @@ describe('README examples', () => {
                 return errs;
               },
             })
-            graduated = undefined as number;
+            public graduated = undefined as number;
           }
 
           const instance = new MyDatesClass();
           assert(instance.valid === false);
           assert(Object.keys(instance.errors).length === 1);
-          assert(instance.errors['born'].length === 1);
+          assert(instance.errors.born.length === 1);
 
           instance.born = 1800;
           assert(instance.valid === false);
-          assert(instance.errors['born'].length === 1);
+          assert(instance.errors.born.length === 1);
 
           instance.born = 1950;
           instance.graduated = 1949;
           assert(instance.valid === false);
-          assert(instance.errors['graduated'].length === 1);
-          assert(instance.errors['graduated'][0].type === 'custom');
+          assert(instance.errors.graduated.length === 1);
+          assert(instance.errors.graduated[0].type === 'custom');
         });
       });
 
@@ -478,30 +478,30 @@ describe('README examples', () => {
 
             // Validate year is within a predefined scope
             @Validate({ provider: Joi.number().strict().integer().min(1900).max(2100).required() })
-            born = undefined as number;
+            public born = undefined as number;
 
             // Validate year is within dynamic scope and throw custom validation errors
             @Validate({ provider: Joi.number().strict().integer().min(Joi.ref('$.born')).max(2100).required() })
-            graduated = undefined as number;
+            public graduated = undefined as number;
           }
 
           const instance = new MyDatesClass();
           assert(instance.valid === false);
           assert(Object.keys(instance.errors).length === 2);
-          assert(instance.errors['born'].length === 1);
-          assert(instance.errors['graduated'].length === 1);
+          assert(instance.errors.born.length === 1);
+          assert(instance.errors.graduated.length === 1);
 
           instance.born = 1800;
           assert(instance.valid === false);
           assert(Object.keys(instance.errors).length === 2);
-          assert(instance.errors['born'].length === 1);
-          assert(instance.errors['graduated'].length === 1);
+          assert(instance.errors.born.length === 1);
+          assert(instance.errors.graduated.length === 1);
 
           instance.born = 1950;
           instance.graduated = 1949;
           assert(instance.valid === false);
           assert(Object.keys(instance.errors).length === 1);
-          assert(instance.errors['graduated'].length === 1);
+          assert(instance.errors.graduated.length === 1);
         });
       });
 
@@ -515,7 +515,7 @@ describe('README examples', () => {
 
             // Validate year is within a predefined scope
             @Validate({ provider: Yup.number().strict().integer().min(1900).max(2100).required() })
-            born = undefined as number;
+            public born = undefined as number;
 
             // Validate year is within dynamic scope and throw custom validation errors
             @Validate({
@@ -526,26 +526,26 @@ describe('README examples', () => {
                 .max(2100)
                 .required(),
             })
-            graduated = undefined as number;
+            public graduated = undefined as number;
           }
 
           const instance = new MyDatesClass();
           assert(instance.valid === false);
           assert(Object.keys(instance.errors).length === 2);
-          assert(instance.errors['born'].length === 1);
-          assert(instance.errors['graduated'].length === 1);
+          assert(instance.errors.born.length === 1);
+          assert(instance.errors.graduated.length === 1);
 
           instance.born = 1800;
           assert(instance.valid === false);
           assert(Object.keys(instance.errors).length === 2);
-          assert(instance.errors['born'].length === 1);
-          assert(instance.errors['graduated'].length === 1);
+          assert(instance.errors.born.length === 1);
+          assert(instance.errors.graduated.length === 1);
 
           instance.born = 1950;
           instance.graduated = 1949;
           assert(instance.valid === false);
           assert(Object.keys(instance.errors).length === 1);
-          assert(instance.errors['graduated'].length === 1);
+          assert(instance.errors.graduated.length === 1);
         });
       });
     });
@@ -566,18 +566,18 @@ describe('README examples', () => {
               Yup.number().strict().integer().min(1900).max(2100).required(),
             ],
           })
-          born = undefined as number;
+          public born = undefined as number;
         }
 
         const instance = new MyDatesClass();
         assert(instance.valid === false);
         assert(Object.keys(instance.errors).length === 1);
-        assert(instance.errors['born'].length === 3);
+        assert(instance.errors.born.length === 3);
 
         instance.born = 1800;
         assert(instance.valid === false);
         assert(Object.keys(instance.errors).length === 1);
-        assert(instance.errors['born'].length === 3);
+        assert(instance.errors.born.length === 3);
       });
     });
 
@@ -590,7 +590,7 @@ describe('README examples', () => {
           }
 
           @Validate({ provider: Yup.number().strict().required() })
-          aNumber = undefined as number;
+          public aNumber = undefined as number;
         }
 
         class MyParentClass extends EnTT {
@@ -600,9 +600,9 @@ describe('README examples', () => {
           }
 
           @Validate({ provider: Yup.boolean().strict().required() })
-          aBoolean = undefined as boolean;
+          public aBoolean = undefined as boolean;
 
-          nested = new MyNestedClass();
+          public nested = new MyNestedClass();
         }
 
         const instance = new MyParentClass();