+++ src/lib/decorators/validate/index.ts
@@ -24,7 +24,7 @@ import { _getDecoratorMetadata } from '../../entt/internals';
  * @param type Type to validate against
  * @param provider Validation provider, supports:
  * - Custom validation function of following shape:
- * - (value, obj) => Error[] | Error | string | boolean
+ * - (value: any, obj?: any) => Error[] | Error | string | boolean
  * - joi:         https://www.npmjs.com/package/joi
  * - joi-browser: https://www.npmjs.com/package/joi-browser
  * - yup:         https://www.npmjs.com/package/yup